fix crash: Note Editor Preview & Read Text
If the 'Read Text' option is enabled, then opening
a card preview (from the note editor) caused a crash

readCardTts tried to access card.q(), which failed as the
card was a preview with unsaved text.

To fix this, create an `EphemeralCard` where .q() works
Extract the functionality from `Note`, as it's only used once

Fixes 14692
