+++ AnkiDroid/src/main/java/com/ichi2/libanki/Note.kt
@@ -330,34 +330,4 @@ class Note : Cloneable {
             return highestClozeId + 1
         }
     }
-
-    fun ephemeralCard(
-        col: Collection,
-        ord: Int = 0,
-        fillEmpty: Boolean = false
-    ): Card {
-        val card = Card(col, null)
-        card.ord = ord
-        card.did = 1
-
-        val nt = notetype
-        val templateIdx = if (nt.type == Consts.MODEL_CLOZE) {
-            0
-        } else {
-            ord
-        }
-        val template = nt.tmpls[templateIdx] as JSONObject
-        template.put("ord", card.ord)
-
-        val output = TemplateManager.TemplateRenderContext.fromCardLayout(
-            this,
-            card,
-            notetype = nt,
-            template = template,
-            fillEmpty = fillEmpty
-        ).render()
-        card.renderOutput = output
-        card.setNote(this)
-        return card
-    }
 }
